Update UsbHostLib to Resolve MSC Disconnect Issue
It was discovered that disconnecting a mass storage device during a
write operation could lead to a hard fault. This was caused by the
driver invoking ctrl_xfer for a device that had already been
disconnected.
To address this issue, usbh_pooling_hubs() has been added to
usbh_umas_reset_disk()."

Change-Id: I7f41e63b7bbcd306bfc354fcd023f87c5231aec7